demo-microservice/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-pubsub
pydantic
